# openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus — Core Business Unified API (Nhom A6)
  version: 1.0.0
  contact:
    name: Nhom A6 Core Business
    email: support@campus.local
  description: |
    Hợp đồng API OpenAPI 3.1 hợp nhất cho nhóm A6 Core Business liên kết với các nhóm khác.
    Bao gồm các cặp đàm phán REST sync:
    - Pair 02: Core Business (A6) ↔ AI Vision (A4)
    - Pair 03: Core Business (A6) ↔ Access Gate (A3) [Core làm Consumer]
    - Pair 10: Access Gate (A3) ↔ Core Business (A6) [Core làm Provider]
servers:
  - url: http://localhost:4010
    description: Prism mock server
  - url: https://api.campus.local
    description: Class network placeholder
...
security:
  - bearerAuth: []
paths:
...
  /vision/detect:
    post:
      operationId: submitVisionDetect
      summary: Yêu cầu phân tích hình ảnh/frame
      description: |
        Gửi hình ảnh (thông qua URL ảnh hoặc chuỗi Base64) để AI Vision thực hiện nhận diện khuôn mặt, biển số xe, đột nhập hoặc các yếu tố an ninh khác.
        Hỗ trợ cả luồng đồng bộ (201 Created) và bất đồng bộ (202 Accepted) phụ thuộc vào tải của hệ thống hoặc yêu cầu của nghiệp vụ.
      tags: [vision]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisionDetectRequest'
            examples:
              detectByUrl:
                value:
                  requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                  cameraId: CAM-001
                  capturedAt: '2026-05-28T15:00:00Z'
                  imageType: URL
                  imageUrl: https://campus.local/images/cam-001/frame-123.jpg
                  locationId: GATE-01
                  metadata:
                    priority: HIGH
              detectByBase64:
                value:
                  requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babd
                  cameraId: CAM-002
                  capturedAt: '2026-05-28T15:01:00Z'
                  imageType: BASE64
                  imageBase64: data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEASABIAAD/2wBDAP...
                  locationId: BLOCK-A
      responses:
        '201':
          description: Phân tích thành công và trả kết quả tức thời (Đồng bộ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectResult'
              examples:
                successMatch:
                  value:
                    requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                    status: SUCCESS
                    detectedAt: '2026-05-28T15:00:02Z'
                    detections:
                      - type: UNKNOWN_PERSON
                        confidence: 0.94
                        boundingBox:
                          x: 120.5
                          y: 85.0
                          width: 45.2
                          height: 110.8
                    processingTimeMs: 245
                    modelVersion: yolov8x-face-v2.1
                    notes: Phát hiện đối tượng chưa đăng ký trong khu vực.
        '202':
          description: Yêu cầu được tiếp nhận và xử lý bất đồng bộ (Async)
          headers:
            Location:
              description: URI để tra cứu kết quả xử lý thông qua GET /vision/detect/{requestId}.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectAsyncResponse'
              examples:
                accepted:
                  value:
                    requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babd
                    status: PROCESSING
                    submittedAt: '2026-05-28T15:01:02Z'
                    estimatedDurationMs: 1500
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vision/detect/{requestId}:
    get:
      operationId: getVisionDetectResult
      summary: Tra cứu kết quả phân tích hình ảnh
      description: Lấy kết quả phân tích hình ảnh hoặc trạng thái xử lý dựa trên requestId đã gửi từ trước.
      tags: [vision]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Trả về kết quả phân tích hình ảnh đã hoàn tất
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectResult'
              examples:
                completedResult:
                  value:
                    requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                    status: SUCCESS
                    detectedAt: '2026-05-28T15:00:02Z'
                    detections:
                      - type: UNKNOWN_PERSON
                        confidence: 0.94
                        boundingBox:
                          x: 120.5
                          y: 85.0
                          width: 45.2
                          height: 110.8
                    processingTimeMs: 245
                    modelVersion: yolov8x-face-v2.1
                    notes: Phát hiện đối tượng chưa đăng ký.
        '202':
          description: Yêu cầu vẫn đang trong hàng đợi hoặc đang xử lý
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectAsyncResponse'
              examples:
                stillProcessing:
                  value:
                    requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babd
                    status: PROCESSING
                    submittedAt: '2026-05-28T15:01:02Z'
                    estimatedDurationMs: 800
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vision/results/recent:
    get:
      operationId: getRecentResults
      summary: Lấy kết quả phân tích gần đây (Phân trang Cursor)
      description: Lấy danh sách lịch sử kết quả phân tích hình ảnh thành công gần đây nhất, áp dụng kỹ thuật Cursor-based Pagination.
      tags: [vision]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách kết quả phân tích phân trang thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectResultPage'
              examples:
                firstPage:
                  value:
                    items:
                      - requestId: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
                        status: SUCCESS
                        detectedAt: '2026-05-28T15:00:02Z'
                        detections:
                          - type: UNKNOWN_PERSON
                            confidence: 0.94
                            boundingBox:
                              x: 120.5
                              y: 85.0
                              width: 45.2
                              height: 110.8
                        processingTimeMs: 245
                        modelVersion: yolov8x-face-v2.1
                        notes: Phát hiện đối tượng chưa đăng ký.
                    nextCursor: 'eyJjcmVhdGVkX2F0IjoiMjAyNi0wNS0yOFQxNTowMDowMFoiLCJpZCI6MTIzfQ=='
                    hasMore: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vision/detections:
    get:
      operationId: listDetectionsHistory
      summary: Lấy lịch sử yêu cầu phân tích (Phân trang Cursor)
      description: Lấy danh sách lịch sử toàn bộ yêu cầu phân tích hình ảnh đã được gửi lên hệ thống (bao gồm cả thành công và thất bại), sử dụng Cursor-based Pagination.
      tags: [vision]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách lịch sử yêu cầu phân tích hình ảnh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDetectResultPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vision/models:
    get:
      operationId: listVisionModels
      summary: Lấy danh sách model AI đang chạy
      description: Trả về danh sách thông tin các model AI nhận diện đang hoạt động để phục vụ audit/debug của Core Business.
      tags: [vision]
      responses:
        '200':
          description: Trả về danh sách model AI đang kích hoạt
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [models]
                properties:
                  models:
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: '#/components/schemas/ModelInfo'
              examples:
                activeModels:
                  value:
                    models:
                      - modelName: yolov8x-face
                        version: v2.1
                        type: FACE_MATCH
                        status: ACTIVE
                      - modelName: yolov8-license-plate
                        version: v1.4
                        type: LICENSE_PLATE
                        status: ACTIVE
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /access/logs/recent:
    get:
      operationId: getAccessLogsRecent
      summary: Lấy danh sách log quẹt thẻ gần đây
      description: |
        Truy xuất danh sách nhật ký quẹt thẻ ra/vào cổng gần đây trên campus.
        Hỗ trợ phân trang dựa trên cursor-based pagination để tối ưu hóa hiệu năng hệ thống.
      tags: [access-gate]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Danh sách log quẹt thẻ được phân trang thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessLogPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /access/logs/{logId}:
    get:
      operationId: getAccessLogById
      summary: Lấy thông tin chi tiết một access log
      description: Truy xuất thông tin chi tiết đầy đủ của một lượt quẹt thẻ bao gồm thông tin chủ thẻ và thiết bị đọc.
      tags: [access-gate]
      parameters:
        - name: logId
          in: path
          required: true
          description: UUID định danh duy nhất của lượt quẹt thẻ.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Truy xuất chi tiết log quẹt thẻ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessLogDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    VisionDetectResult:
      type: object
      additionalProperties: false
      required: [requestId, status, detectedAt, detections]
      properties:
        requestId:
          type: string
          format: uuid
          description: Khớp với requestId trong yêu cầu gửi lên.
          example: 0196fb3d-4ad7-7d1e-9f49-5d5148d2babc
        status:
          type: string
          enum: [SUCCESS, FAILED]
          description: Trạng thái xử lý phân tích hình ảnh.
          example: SUCCESS
        detectedAt:
          type: string
          format: date-time
          description: Thời điểm AI hoàn thành việc phân tích hình ảnh.
          example: '2026-05-28T15:00:02Z'
        detections:
          type: array
          minItems: 0
          maxItems: 100
          description: Danh sách các đối tượng nhận dạng được.
          items:
            $ref: '#/components/schemas/DetectionItem'
        processingTimeMs:
          type: integer
          minimum: 0
          description: Thời gian xử lý ảnh tính bằng mili-giây.
          example: 245
        modelVersion:
          type: string
          minLength: 1
          maxLength: 80
          description: Tên và phiên bản model AI phân tích thực tế.
          example: yolov8x-face-v2.1
        notes:
          type: [string, 'null']
          maxLength: 500
          description: Ghi chú hoặc thông điệp chi tiết bổ sung.
          example: Phát hiện đối tượng chưa đăng ký trong khu vực.
...